+++ src/iStrategy.ts
