+++ types/iStrategy.d.ts
