+++ types/index.d.ts
@@ -1,8 +1,8 @@
-export * from "./Decrator";
-export * from "./IStrategy";
-export * from "./Watch";
-export * from "./Ast";
-export * from "./Utils";
+export * from "./decrator";
+export * from "./iStrategy";
+export * from "./watch";
+export * from "./ast";
+export * from "./utils";
 export * from "./parse";
 export * from "./loadObject/scan";
 export * from "./loadObject/require";
