+++ types/uvuStrategy.d.ts
@@ -1,4 +1,4 @@
-import IStrategy from './IStrategy';
+import IStrategy from './iStrategy';
 export default class UvuStrategy implements IStrategy {
     test: any;
     testcase(name: string): any;
