+++ types/watch.d.ts
@@ -1,2 +1,2 @@
-import Context from "./Context";
+import Context from "./context";
 export declare function WatchFiles(testFiles: string[], context: Context): void;
